+++ changelog.md
@@ -1,2 +1,7 @@
 2022-08-09: v1.0.0
 - Inital version
+
+2022-08-17: v1.1.0
+- Better user resolve on mentions
+- Allow bot mention to run commands
+- Better channel / role resolve on mentions